python/offline/pyproject.toml
[project]
name = "aoi360-pipeline"
version = "0.1.0"
description = "Offline AOI generation pipeline for 360 eye tracking"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate",
    "numpy",
    "opencv-python",
    "pandas",
    "pillow",
    "torch",
    "torchvision",
    "tqdm",
    "transformers",
]

[project.scripts]
aoi360-extract-frames = "aoi360_pipeline.frame_extraction:main"
aoi360-detect-grounding-dino = "aoi360_pipeline.grounding_dino:main"
aoi360-build-aoi-map = "aoi360_pipeline.aoi_map_builder:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
